+++ internal/xsql/rows_test.go
@@ -7,7 +7,11 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/common"
 )
+
+var _ common.Rows = (*singleRow)(nil)
 
 func TestRowByAstPlan(t *testing.T) {
 	t.Run("CreateRow", func(t *testing.T) {
@@ -27,7 +31,7 @@ func TestRowByAstPlan(t *testing.T) {
 
 func TestSingleRow_Columns(t *testing.T) {
 	row := rowByAstPlan("ast", "plan")
-	columns := row.Columns()
+	columns := row.Columns(t.Context())
 
 	require.Len(t, columns, 2)
 	require.Equal(t, "Ast", columns[0])
@@ -36,18 +40,20 @@ func TestSingleRow_Columns(t *testing.T) {
 
 func TestSingleRow_Close(t *testing.T) {
 	row := rowByAstPlan("ast", "plan")
-	err := row.Close()
+	err := row.Close(t.Context())
 	require.NoError(t, err)
 }
 
 func TestSingleRow_Next(t *testing.T) {
+	ctx := t.Context()
+
 	t.Run("FirstRead", func(t *testing.T) {
 		ast := "test ast"
 		plan := "test plan"
 		row := rowByAstPlan(ast, plan)
 
 		dst := make([]driver.Value, 2)
-		err := row.Next(dst)
+		err := row.Next(ctx, dst)
 
 		require.NoError(t, err)
 		require.Equal(t, ast, dst[0])
@@ -59,10 +65,10 @@ func TestSingleRow_Next(t *testing.T) {
 		row := rowByAstPlan("ast", "plan")
 
 		dst := make([]driver.Value, 2)
-		err := row.Next(dst)
+		err := row.Next(ctx, dst)
 		require.NoError(t, err)
 
-		err = row.Next(dst)
+		err = row.Next(ctx, dst)
 		require.ErrorIs(t, err, io.EOF)
 	})
 
@@ -73,7 +79,7 @@ func TestSingleRow_Next(t *testing.T) {
 		}
 
 		dst := make([]driver.Value, 2)
-		err := row.Next(dst)
+		err := row.Next(ctx, dst)
 		require.ErrorIs(t, err, io.EOF)
 	})
 
@@ -84,7 +90,7 @@ func TestSingleRow_Next(t *testing.T) {
 		}
 
 		dst := make([]driver.Value, 0)
-		err := row.Next(dst)
+		err := row.Next(ctx, dst)
 		require.NoError(t, err)
 		require.True(t, row.readAll)
 	})